CAT Auto Pulish for Windows Release Messages - 20190916182624 (#1124)
* Microsoft Defender ATP: amend broken link

The link "Get started" at the end of the page "Microsoft Defender
Advanced Threat Protection" points to a non-existing page. It may have
been renamed and the name "Get started" appropriated as one of the
expandable sections in the left hand side index/TOC section.

The suggested change is to replace the broken link `get-started.md`
with `minimum-requirements.md` found at the same directory level.

Broken link:
https://github.com/MicrosoftDocs/windows-itpro-docs/blob/master/windows/security/threat-protection/microsoft-defender-atp/get-started.md

Replacement link:
https://github.com/MicrosoftDocs/windows-itpro-docs/blob/master/windows/security/threat-protection/microsoft-defender-atp/minimum-requirements.md

Thanks to xNiklasJern for discovering the broken link.

Closes #4832

* Added update to run the task

Updated the document as suggested by the user to update the XML file created by GP preference. 

Problem: https://github.com/MicrosoftDocs/windows-itpro-docs/issues/4797

* Update windows/security/threat-protection/microsoft-defender-atp/configure-endpoints-gp.md

Co-Authored-By: JohanFreelancer9 <48568725+JohanFreelancer9@users.noreply.github.com>

* Update hello-faq.md: English grammar improvement

English grammar improvement is needed in one of the first sentences of
this document page: " but not date at this time."

Thanks to robhurt for noticing and reporting this malformed sentence.

Closes #4932

* MBAM 2.5 Service releases: add May 2019 to list

This is yet another page where the May 2019 Service release is missing
from the list of service releases of Microsoft Desktop Optimization Pack
(MDOP).

Proposed change: add May 2019 Service release at the end of the list.

Thanks to philschwan for noticing and reporting this detail.

Closes #4944

* fixed spelling error

* fixed spelling error

* CAT Auto Pulish for Windows Release Messages - CAT_AutoPublish_20190916173707 (#1123)