Auditing: add MD code block to table keywords
Description:
This PR aims to block keywords and values from being translated to
another language, keeping the values and keywords intact, in response
to the windows-itpro-docs issue ticket #4995.

Proposed changes:

- Surround Logon Type values and Logon Title keywords with MD code
  block markers (back ticks) to keep them from being translated by MT.
- Add XML indentation to the XML for the layout to be shown properly.

Thanks to @takondo for pointing out the uselessness in translating
these parts of the table when the document is machine translated.

Ref. issue ticket #4995

(The ticket can be closed when this commit is successfully migrated
  and shown to be effective against machine translation.)