WDAC/Recommended block rules: Add notes and link
Description:

As agreed on in issue ticket #3642 (Ambiguity), this PR adds a note
about creating two diff policies or merging them in a broad policy,
as well as adding a link to how to merge WDAC policies.

As recommended by Justin Hall, Windows Server 2019 is also added to the
"Applies to" list at the top of this document.

Also, thanks to Air-Git for requesting this clarification.

issue ticket closure or reference:

Closes #3642