MDATP/ASR rules: Restore broken link
Description:
As described in issue ticket #5105, the link to "adding exclusions" for
Attack Surface Reduction Rules is broken and causes a HTTP 404 error.

Thanks to redog for reporting the broken link & 404 error.

The cause of this link breaking is more or less documented on the
Github source page for "Reduce attack surfaces with attack surface
reduction rules" with the latest commit ca9b815
(moved attack surface topics to mdatp dir) from August 21, 2019.

Proposed changes:
- insert the correct new page link, pointing to the Github page
- correct an old typo, securty -> security

issue ticket closure or reference:

Closes #5105