Advanced hunting query language: Kusto query typos
Description:

According to the report in issue ticket #5457
(Error in example code Try your first query),
the sample query fails because of the quote apostrophe at the end of
the last line in the query. As can be seen in the complementary image,
the single quote or apostrophe is unexpected in that position. This can
also be verified by using query language editors, marking the quote as
an unexpected error in the text.

Thanks to Gene Cupstid for reporting the typo in this training sample.

Changes proposed:

- remove 2 occurrences of the unwanted quote/apostrophe
- add "kusto" as recommended MD syntax highlighting tag,
  as recommended in Contribute/Contribute/how-to-write-use-markdown.md
- add MarkDown compatibility spacing in the MD Note quote markers

issue ticket closure or reference:

Closes #5457