VDI/WDA Deployment guide: 404 URL link correction
Description:

As described in issue ticket #5485, the link in the last line of this
document points to an "Error 404 - Page Not Found" page.

The working page can be reached by removing the last directory name in
the URL (only works when clicking the correct link, not if you edit the
wrong URL to get the correct URL).

Thanks to amirsc3 for reporting the issue + providing the correct URL.

Changes proposed:
- remove the incorrect directory name at the end of the URL
- add a few MD compatibility spaces in the MD Notes
- remove some redundant whitespace between text sections

issue ticket closure or reference:

Closes #5485