Update mac-preferences.md
The format of the .plist for Intune was wrong before. The correct format doesn't have anything wrapped in <xml> or <dict> or <plist>. This needes to be edited as there was an ICM caused due to customer confusion in the documentation. Only key value pairs are needed and required for the correct format