WDAC policy/Rule option 18: clarify ambiguity
Description:

According to the latest feedback from Air-Git in issue ticket #5462
(Option 18 is ambiguous), in the last part of the Rule option
Description, the wording is now precise, but not accurate.

An accurate statement is as follows:
"[...] for any FileRule that allows a file based on FilePath."

Thanks to Air-Git for the continued valuable feedback to MS Docs
articles in need of improvement and suggested solutions.

Proposed change:
- Adjust the end part of the sentence to be accurate.

issue ticket closure or reference:

Closes #5462