MD-ATP for Linux: typos & code block corrections
Description:

As reported in issue ticket #6443
(the copy paste fields on this doc are incorrect
 and can cause errors/confusion), there are 3 lines incorrectly added
into the copy-paste blocks in this deployment description.

There is also at least 1 copy-paste block in need of moving the
actual commands away from the console output and out into their own
boxes or monospace command line notation.

Thanks to bled1982 for reporting this issue.

Changes proposed:
- Remove 1 line containing "[your organization identifier]"
- Remove 2 occurrences of an unwarranted "1" character line
- Split the 'ls -l' command from the console output
- Split the 'unzip' command from the console output
- Adjust code block indent for 3 double-indented blocks

Ticket closure or reference:

Closes #6443