MD-ATP/Exploit Protection: corrections & updates
Description:

As reported in issue ticket #6531 (Redirect Exploit protection
"Evaluation Package" link directly to right site), the aka.ms link to
the Evaluation Package points to the Windows Defender test ground page.
The suggestion is to point the link directly to the Exploit Protection
(EP) page instead.

Thanks to beerisgood for reporting this issue.

Additional changes: Link corrections and updates to counteract the fact
that some of the old technet links do not lead directly to the correct
topic pages, so it is better to insert traceable direct links instead.

I have not found any formatting worth correcting on this page, so the
various outdated or bad links will be the main focus in this PR.

Changes proposed:
- Replace the aka.ms/mp7z2w link with the direct EP download page link
- Replace technet/msdn links with current docs.microsoft.com page links
- Replace 1 technet link with support.microsoft.com/help
- Remove bad go.microsoft.com/fwlink (Windows 10 edition comparison PDF)
- Insert direct link to local neighboring page for MD-ATP (same folder)

Ticket closure or reference:

Closes #6531