Update Event Id 4624
In the Network information section, the workstation name is mentioned with 'machine name from which logon attempt was performed.' that has been modified to "machine name to which logon attempt was performed." (please notice 'to' instead of 'from'). The workstation name has to be the computer name of the target computer not the source computer. Please check the issue reported on https://github.com/MicrosoftDocs/windows-itpro-docs/issues/6438, thanks.