Security/Threat protection: password length values
Description:

As requested by Program Manager Robert Durff (MSRobertD) in issue
ticket #6856 (Bug: Password length value range is inaccurate.), the
upper value for the supported values for password length should be 20
instead of only 14, verified in preliminary field testing of the GPO
Password Policy, described on this page.

The actual upper limit may very well be higher, but 20 is a reasonable
value to be used for now, until someone documents the need for higher
accuracy in the documentation of this value for the GPO Password Policy.

Changes proposed:
- Replace 14 with 20 in both occurrences of 14 as the upper value
- Convert Note text in line 83 to a MarkDown Note blob (MS codestyle)
- Whitespace adjustments:
    - Normalize bullet point list spacing to 1 (codestyle) (3 lines)
    - Remove redundant end-of-line spacing (8 lines)

Ticket closure or reference:

Closes #6856