MD-ATP/Exclusions: avoid translation of keywords
Description:

As recommended by Tina McNaboe in issue ticket #7021 (Doubt), I suggest
encapsulating all OS strings and keywords in MarkDown back ticks to
avoid automatic translation of text that should never be translated, as
well as keeping the strings as coherent and unmodified as possible.

To paraphrase:
While the engineers are working on a fix to avoid overlocalization like
this on machine translated articles, it may be a few months before
it's fixed. In the meantime, it would be helpful to encapsulate the
non-localizable strings in the English article.

It is also a common problem that Machine Translation adds spacing where
it should not be, even if the surrounding text is kept untranslated.
This type of issue tends to be less easy to discover for non-English
readers who rely more on the translated pages than on the /en-us/ pages.

This PR aims to reduce these issues, at least for this page, to keep it
as correct as possible, even when filtered through Machine Translation.

Changes proposed:
- Encapsulate listed strings, filenames and extensions in MD back ticks
- Make sure that the encapsulated strings don't keep extra backslashes

Additional corrections, since we are already editing this page:
- Replace TechNet URLs with their permanent docs.microsoft.com redirects
- Replace MSDN URLs with their permanent docs.microsoft.com redirects
- Reduce a docs.microsoft.com link to internal tag link (same page)
- Adjust "Currentcontrolset" to CurrentControlSet (3 occurrences)

Ticket closure or reference:

Ref.  #7021 (keep open until changes migrate to localized pages)