Policy GUID needs clarity
I think this doc needs some clarity around the Policy GUID value as it is not clear not to include the curly brackets. Although this is shown in the image it can be quite small and difficult to see. I've added a note under the section but there might be a better approach.